+++ R/Maze.R
@@ -117,8 +117,8 @@
 #' policy(maze_solved)
 #'
 #' # show the utilities and optimal actions organized in the maze layout (like in the AIMA textbook)
-#' matrix(policy(maze_solved)[[1]]$U, nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
-#' matrix(policy(maze_solved)[[1]]$action, nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
+#' matrix(policy(maze_solved)$U, nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
+#' matrix(policy(maze_solved)$action, nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
 #' # Note: the optimal actions for the states with a utility of 0 are artefacts and should be ignored.
 #'
 #' # plot the solution as an image plot. Darker colors represent larger
@@ -144,7 +144,7 @@ plot_Maze_solution <-
            states = TRUE,
            main = "Russell's 4x3 Maze",
            ...) {
-    pi <- policy(x)[[epoch]]
+    pi <- policy(x, drop = FALSE)[[epoch]]
     U <- pi$U
     U[absorbing_states(x)] <- NA