+++ R/policy.R
@@ -7,21 +7,25 @@
 #' converged solution is produced.
 #' For a POMDP, the policy is a data.frame consisting of:
 #'
-#' * Part 1: The value function with one column per state (alpha vectors).
-#' * Part 2: The last column contains the prescribed action.
+#' * Part 1: The alpha vectors for the belief states (defines also the 
+#'            utility of the belief). The columns have 
+#'            the names of states.
+#' * Part 2: The last column named `action` contains the prescribed action.
 #'
-#' For an MDP, the policy is a data.frame consisting of:
+#' For an MDP, the policy is a data.frame with columns for:
 #'
-#' * The state
-#' * The state's discounted expected utility U if the policy is followed
-#' * The prescribed action
+#' * `state`: The state.
+#' * `U`: The state's value (discounted expected utility U) if the policy 
+#'    is followed
+#' * `action`: The prescribed action.
 #'
 #' @family policy
 #'
 #' @param x A solved [POMDP] or [MDP] object.
-#' @param alpha logical; include the parameters of the alpha vector defining the segment (POMDP only).
-#' @param action logical; include the action for that segment (POMDP only).
-#' @return A list with the policy for each epoch.
+#' @param drop logical; drop the list for converged, epoch-independent policies.
+#' @return A list with the policy for each epoch. Converged policies
+#'  have only one element. If `drop = TRUE` then the policy is returned 
+#'  without a list.
 #' @author Michael Hahsler
 #' @keywords graphs
 #' @examples
@@ -36,11 +40,13 @@
 #' plot_value_function(sol)
 #'
 #' # Finite horizon (we use incremental pruning because grid does not converge)
-#' sol <- solve_POMDP(model = Tiger, method = "incprune", horizon = 3, discount = 1)
+#' sol <- solve_POMDP(model = Tiger, method = "incprune", 
+#'   horizon = 3, discount = 1)
 #' sol
 #'
 #' policy(sol)
-#' # Note: We see that it is initially better to listen till we make a decision in the final epoch.
+#' # Note: We see that it is initially better to listen till we make 
+#' #       a decision in the final epoch.
 #'
 #' # MDP policy
 #' data(Maze)
@@ -49,33 +55,36 @@
 #'
 #' policy(sol)
 #' @export
-policy <- function(x, alpha = TRUE, action = TRUE) {
+policy <- function(x, drop = TRUE) {
   UseMethod("policy")
 }
 
 #' @export
-policy.MDP <- function(x, alpha = TRUE, action = TRUE) {
+policy.MDP <- function(x, drop = TRUE) {
   is_solved_MDP(x, stop = TRUE)
-  x$solution$policy
+  
+  policy <- x$solution$policy
+  if(drop && length(policy) == 1)
+    policy <- policy[[1]]
+  
+  policy
 }
 
 #' @export
-policy.POMDP <- function(x, alpha = TRUE, action = TRUE) {
+policy.POMDP <- function(x, drop = TRUE) {
   is_solved_POMDP(x, stop = TRUE)
   n <- length(x$solution$pg)
   
   policy <- vector("list", n)
   for (i in seq_len(n)) {
-    if (alpha && !action)
-      policy[[i]] <- x$solution$alpha[[i]]
-    else if (!alpha && action)
-      policy[[i]] <-
-        x$solution$pg[[i]][, "action", drop = FALSE]
-    else
       policy[[i]] <- cbind(x$solution$alpha[[i]],
                            x$solution$pg[[i]][, "action", drop = FALSE])
   }
-  return(policy)
+  
+  if(drop && length(policy) == 1)
+    policy <- policy[[1]]
+  
+  policy
 }