+++ R/solve_MDP.R
@@ -6,7 +6,8 @@
 #' methods based on reinforcement learning techniques to solve finite
 #' state space MDPs.
 #'
-#' Implemented are the following dynamic programming methods:
+#' Implemented are the following dynamic programming methods (following
+#' Russell and Norvig, 2010):
 #'
 #' * **Modified Policy Iteration**
 #' starts with a random policy and iteratively performs
@@ -25,8 +26,9 @@
 #' is calculated from the value function. Value iteration can be seen as
 #' policy iteration with truncated policy evaluation.
 #'
-#' Implemented are the following temporal difference control methods. Note
-#' that the MDP transition and reward models are only used to simulate
+#' Implemented are the following temporal difference control methods
+#' following Sutton and Barto (2020).
+#' Note that the MDP transition and reward models are only used to simulate
 #' the environment for these reinforcement learning methods.
 #' The algorithms use a step size parameter \eqn{\alpha} (learning rate) for the
 #' updates and the exploration parameter \eqn{\epsilon} for
@@ -71,6 +73,13 @@
 #'   - `iterations` number of iterations to convergence (infinite-horizon only)
 #'
 #' @author Michael Hahsler
+#' @references 
+#' Russell, S., Norvig, P. (2021). Artificial Intelligence: A Modern Approach. 
+#' Fourth edition. Prentice Hall.
+#' 
+#' Sutton, R. S., Barto, A. G. (2020). Reinforcement Learning: An Introduction. 
+#' Second edition. The MIT Press.
+#' 
 #' @examples
 #' data(Maze)
 #' Maze