+++ tests/testthat/test-solve_MDP.R
@@ -10,7 +10,7 @@ methods <- c("value_iteration", "policy_iteration", "q_learning", "sarsa",
 
 for (m in methods) {
   sol <- solve_MDP(Maze, method = m)
-  pol <- policy(sol)[[1]]
+  pol <- policy(sol)
   expect_identical(dim(pol), c(length(Maze$states), 3L))
   
   #check_and_fix_MDP(sol)